+++ src/crags/resolvers/sectors.resolver.ts
@@ -7,7 +7,13 @@ import {
   Query,
 } from '@nestjs/graphql';
 import { Roles } from '../../auth/decorators/roles.decorator';
-import { UseInterceptors, UseFilters, UseGuards } from '@nestjs/common';
+import {
+  UseInterceptors,
+  UseFilters,
+  UseGuards,
+  BadRequestException,
+  ForbiddenException,
+} from '@nestjs/common';
 import { AuditInterceptor } from '../../audit/interceptors/audit.interceptor';
 import { NotFoundFilter } from '../filters/not-found.filter';
 import { Sector } from '../entities/sector.entity';
@@ -15,42 +21,86 @@ import { SectorsService } from '../services/sectors.service';
 import { CreateSectorInput } from '../dtos/create-sector.input';
 import { UpdateSectorInput } from '../dtos/update-sector.input';
 import { Route } from '../entities/route.entity';
-import { Loader } from '../../core/interceptors/data-loader.interceptor';
-import { SectorRoutesLoader } from '../loaders/sector-routes.loader';
-import DataLoader from 'dataloader';
 import { AllowAny } from '../../auth/decorators/allow-any.decorator';
 import { UserAuthGuard } from '../../auth/guards/user-auth.guard';
 import { ForeignKeyConstraintFilter } from '../filters/foreign-key-constraint.filter';
+import { CurrentUser } from '../../auth/decorators/current-user.decorator';
+import { User } from '../../users/entities/user.entity';
+import { RoutesService } from '../services/routes.service';
+import { NotificationService } from '../../notification/services/notification.service';
 
 @Resolver(() => Sector)
 export class SectorsResolver {
-  constructor(private sectorsService: SectorsService) {}
+  constructor(
+    private sectorsService: SectorsService,
+    private routesService: RoutesService,
+    private notificationService: NotificationService,
+  ) {}
+
+  /* QUERIES */
 
   @Query(() => Sector)
   @UseFilters(NotFoundFilter)
   @AllowAny()
   @UseGuards(UserAuthGuard)
-  sector(@Args('id') id: string): Promise<Sector> {
-    return this.sectorsService.findOneById(id);
+  sector(@Args('id') id: string, @CurrentUser() user: User): Promise<Sector> {
+    return this.sectorsService.findOne({ id: id, user: user });
   }
 
+  /* MUTATIONS */
+
   @Mutation(() => Sector)
-  @Roles('admin')
+  @UseGuards(UserAuthGuard)
   @UseInterceptors(AuditInterceptor)
   @UseFilters(NotFoundFilter)
   async createSector(
     @Args('input', { type: () => CreateSectorInput }) input: CreateSectorInput,
+    @CurrentUser() user: User,
   ): Promise<Sector> {
-    return this.sectorsService.create(input);
+    if (!user.isAdmin() && input.publishStatus == 'published') {
+      throw new BadRequestException();
+    }
+    return this.sectorsService.create(input, user);
   }
 
   @Mutation(() => Sector)
-  @Roles('admin')
-  @UseInterceptors(AuditInterceptor)
+  @UseGuards(UserAuthGuard)
   @UseFilters(NotFoundFilter)
+  @UseInterceptors(AuditInterceptor)
   async updateSector(
     @Args('input', { type: () => UpdateSectorInput }) input: UpdateSectorInput,
+    @CurrentUser() user: User,
   ): Promise<Sector> {
+    const sector = await this.sectorsService.findOne({
+      id: input.id,
+      user,
+    });
+
+    if (!user.isAdmin() && sector.publishStatus != 'draft') {
+      throw new ForbiddenException();
+    }
+
+    if (!user.isAdmin() && input.publishStatus == 'published') {
+      throw new BadRequestException('publish_status_unavailable_to_user');
+    }
+
+    const crag = await sector.crag;
+    if (
+      input.publishStatus != null &&
+      crag.publishStatus < input.publishStatus
+    ) {
+      throw new BadRequestException('publish_status_incompatible_with_crag');
+    }
+
+    if (sector.publishStatus == 'in_review' && input.publishStatus == 'draft') {
+      this.notificationService.contributionRejection(
+        { sector: sector },
+        await sector.user,
+        user,
+        input.rejectionMessage,
+      );
+    }
+
     return this.sectorsService.update(input);
   }
 
@@ -66,20 +116,36 @@ export class SectorsResolver {
   }
 
   @Mutation(() => Boolean)
-  @Roles('admin')
+  @UseGuards(UserAuthGuard)
   @UseInterceptors(AuditInterceptor)
   @UseFilters(NotFoundFilter, ForeignKeyConstraintFilter)
-  async deleteSector(@Args('id') id: string): Promise<boolean> {
+  async deleteSector(
+    @Args('id') id: string,
+    @CurrentUser() user: User,
+  ): Promise<boolean> {
+    const sector = await this.sectorsService.findOne({
+      id: id,
+      user,
+    });
+
+    if (!user.isAdmin() && sector.publishStatus != 'draft') {
+      throw new ForbiddenException();
+    }
+
     return this.sectorsService.delete(id);
   }
+
+  /* FIELDS */
 
   @ResolveField('routes', () => [Route])
   async getRoutes(
     @Parent() sector: Sector,
-    @Loader(SectorRoutesLoader)
-    loader: DataLoader<Route['id'], Route[]>,
+    @CurrentUser() user: User,
   ): Promise<Route[]> {
-    return loader.load(sector.id);
+    return this.routesService.find({
+      sectorId: sector.id,
+      user,
+    });
   }
 
   @ResolveField('bouldersOnly', () => Boolean)
